chore: reorganized server file structure
