+++ src/server/domains/sms/smsServices.ts
@@ -1,4 +1,4 @@
-import type { SmsResponse } from "../../types/sms.types";
+import type { SmsResponse } from "./smsTypes.ts";
 
 export async function sendSMS(to: string, message: string): Promise<SmsResponse> {
